+++ ext/Server/BotSpawner.lua
@@ -28,6 +28,8 @@ function BotSpawner:RegisterVars()
 	self._PlayerUpdateTimer = 0.0
 	self._FirstSpawnInLevel = true
 	self._FirstSpawnDelay = Registry.BOT_SPAWN.FIRST_SPAWN_DELAY
+	self._DelayDirectSpawn = Registry.BOT_SPAWN.DELAY_DIRECT_SPAWN
+	self._NrOfPlayers = 0
 	self._UpdateActive = false
 	---@type SpawnSet[]
 	self._SpawnSets = {}
@@ -54,7 +56,7 @@ function BotSpawner:OnLevelLoaded(p_Round)
 
 	if (Config.TeamSwitchMode == TeamSwitchModes.SwitchForRoundTwo and p_Round ~= self._LastRound) or
 		(Config.TeamSwitchMode == TeamSwitchModes.AlwaysSwitchTeams) or
-		(Config.TeamSwitchMode == TeamSwitchModes.SwitchTeamsRandomly and MathUtils:GetRandom(1, 100) >= 50)
+		(Config.TeamSwitchMode == TeamSwitchModes.SwitchTeamsRandomly and m_Utilities:CheckProbablity(50))
 	then
 		m_Logger:Write("switch teams")
 		self:_SwitchTeams()
@@ -69,7 +71,9 @@ function BotSpawner:OnLevelDestroy()
 	self._UpdateActive = false
 	self._FirstSpawnInLevel = true
 	self._FirstSpawnDelay = Registry.BOT_SPAWN.FIRST_SPAWN_DELAY
+	self._DelayDirectSpawn = Registry.BOT_SPAWN.DELAY_DIRECT_SPAWN
 	self._PlayerUpdateTimer = 0.0
+	self._NrOfPlayers = 0
 end
 
 -- =============================================
@@ -94,6 +98,10 @@ function BotSpawner:OnUpdateManagerUpdate(p_DeltaTime, p_UpdatePass)
 			self._FirstSpawnDelay = self._FirstSpawnDelay - p_DeltaTime
 		end
 	else
+		-- count below 0 for post-delay-reactions
+		if self._DelayDirectSpawn > -10 and Globals.IsInputAllowed and self._NrOfPlayers > 0 then
+			self._DelayDirectSpawn = self._DelayDirectSpawn - p_DeltaTime
+		end
 		self._PlayerUpdateTimer = self._PlayerUpdateTimer + p_DeltaTime
 
 		if self._PlayerUpdateTimer > 2.0 then
@@ -117,7 +125,7 @@ function BotSpawner:OnUpdateManagerUpdate(p_DeltaTime, p_UpdatePass)
 
 			if Globals.SpawnMode ~= SpawnModes.manual then
 				-- Garbage-collection of unwanted bots
-				m_BotManager:DestroyDisabledBots()
+				m_BotManager:DestroyDisabledBots(self._DelayDirectSpawn > -5.0)
 				m_BotManager:RefreshTables()
 			end
 		end
@@ -134,6 +142,7 @@ function BotSpawner:OnUpdateManagerUpdate(p_DeltaTime, p_UpdatePass)
 				for j, l_BotNameToIgnore in pairs(Globals.IgnoreBotNames) do
 					if l_BotNameToIgnore == l_PlayerNameToKick then
 						table.remove(Globals.IgnoreBotNames, j)
+						break
 					end
 				end
 
@@ -320,6 +329,14 @@ end
 -- =============================================
 -- Public Functions
 -- =============================================
+function BotSpawner:UpdateBotNames()
+	self._SpawnSets = {}
+	local s_TimeToWait = (#m_BotManager:GetBots() * Registry.BOT.BOT_DESTORY_DELAY) + 1.0
+	m_BotManager:DestroyAll()
+	m_BotManager:RefreshTables()
+	self._FirstSpawnInLevel = true
+	self._FirstSpawnDelay = s_TimeToWait
+end
 
 function BotSpawner:UpdateBotAmountAndTeam()
 	-- Keep Slot for next player.
@@ -349,11 +366,18 @@ function BotSpawner:UpdateBotAmountAndTeam()
 	local s_CountBots = {}
 	local s_TargetTeamCount = {}
 
+	local s_BotsToDelay = {}
+
 	for i = 1, Globals.NrOfTeams do
 		s_CountPlayers[i] = 0
-		s_CountBots[i] = m_BotManager:GetActiveBotCount(i)
+		s_BotsToDelay[i] = 0
+		s_CountBots[i] = m_BotManager:GetActiveBotCount(i) -- + m_BotManager:GetInactiveBotCount()
 		s_TargetTeamCount[i] = 0
 		local s_TempPlayers = PlayerManager:GetPlayersByTeam(i)
+
+		if self._DelayDirectSpawn > -5.0 then
+			s_BotsToDelay[i] = #g_GameDirector:GetSpawnableVehicle(i)
+		end
 
 		for _, l_Player in pairs(s_TempPlayers) do
 			if not m_Utilities:isBot(l_Player) then
@@ -361,9 +385,11 @@ function BotSpawner:UpdateBotAmountAndTeam()
 			end
 		end
 
-		s_TeamCount[i] = s_CountBots[i] + s_CountPlayers[i]
+		s_TeamCount[i] = s_CountBots[i] + s_CountPlayers[i] + s_BotsToDelay[i]
 		s_PlayerCount = s_PlayerCount + s_CountPlayers[i]
 	end
+
+	self._NrOfPlayers = s_PlayerCount
 
 	-- Kill and destroy bots, if no player left.
 	if s_PlayerCount == 0 then
@@ -627,7 +653,7 @@ end
 ---@param p_Spacing number
 function BotSpawner:SpawnBotRow(p_Player, p_Length, p_Spacing)
 	for i = 1, p_Length do
-		local s_Name = m_BotCreator:GetNextBotName(self:_GetSpawnBotKit())
+		local s_Name = m_BotCreator:GetNextBotName(self:_GetSpawnBotKit(), m_BotManager:GetBotTeam())
 
 		if s_Name ~= nil then
 			local s_Transform = LinearTransform()
@@ -648,7 +674,7 @@ end
 ---@param p_Height integer
 function BotSpawner:SpawnBotTower(p_Player, p_Height)
 	for i = 1, p_Height do
-		local s_Name = m_BotCreator:GetNextBotName(self:_GetSpawnBotKit())
+		local s_Name = m_BotCreator:GetNextBotName(self:_GetSpawnBotKit(), m_BotManager:GetBotTeam())
 
 		if s_Name ~= nil then
 			local s_Yaw = p_Player.input.authoritativeAimingYaw
@@ -675,7 +701,7 @@ end
 function BotSpawner:SpawnBotGrid(p_Player, p_Rows, p_Columns, p_Spacing)
 	for i = 1, p_Rows do
 		for j = 1, p_Columns do
-			local s_Name = m_BotCreator:GetNextBotName(self:_GetSpawnBotKit())
+			local s_Name = m_BotCreator:GetNextBotName(self:_GetSpawnBotKit(), m_BotManager:GetBotTeam())
 
 			if s_Name ~= nil then
 				local s_Yaw = p_Player.input.authoritativeAimingYaw
@@ -1055,8 +1081,19 @@ function BotSpawner:_SpawnSingleWayBot(p_Player, p_UseRandomWay, p_ActiveWayInde
 		s_IsRespawn = true
 		s_Name = p_ExistingBot.m_Name
 		s_TeamId = p_ExistingBot.m_Player.teamId
-	else
-		s_Name = m_BotCreator:GetNextBotName(self:_GetSpawnBotKit())
+	elseif Registry.BOT_SPAWN.KEEP_BOTS_ON_NEW_ROUND and m_BotManager:GetInactiveBotCount(s_TeamId) > 0 then
+		local s_Bot = m_BotManager:GetInactiveBot(s_TeamId)
+		if s_Bot then
+			p_ExistingBot = s_Bot
+			s_IsRespawn = true
+			s_Name = p_ExistingBot.m_Name
+			s_TeamId = p_ExistingBot.m_Player.teamId
+		end
+	end
+
+	-- only new bot, if no respawn
+	if not s_IsRespawn or not p_ExistingBot then
+		s_Name = m_BotCreator:GetNextBotName(self:_GetSpawnBotKit(), s_TeamId)
 	end
 
 	local s_SquadId = self:_GetSquadToJoin(s_TeamId)
@@ -1119,7 +1156,10 @@ function BotSpawner:_SpawnSingleWayBot(p_Player, p_UseRandomWay, p_ActiveWayInde
 							break
 						end
 					end
+				elseif s_SpawnPoint == "SpawnInGunship" then
+					s_SpawnEntity = g_GameDirector:GetGunship(s_TeamId)
 				end
+
 
 				if s_IsRespawn and p_ExistingBot then
 					p_ExistingBot:SetVarsWay(nil, true, 0, 0, false)
@@ -1289,12 +1329,27 @@ function BotSpawner:_GetSpawnPoint(p_TeamId, p_SquadId)
 	local s_MaximumTrys = 100
 	local s_TrysDone = 0
 
-	if Config.UseVehicles and #g_GameDirector:GetSpawnableVehicle(p_TeamId) > 0 then
+	if Config.UseVehicles and self._DelayDirectSpawn <= 0.0 and #g_GameDirector:GetSpawnableVehicle(p_TeamId) > 0 then
 		return "SpawnAtVehicle"
 	end
 
 	if Config.AABots and #g_GameDirector:GetStationaryAas(p_TeamId) > 0 then
 		return "SpawnInAa"
+	end
+
+	-- Gunships disabled for now! TODO: enable gunship again, once server-cameras are supported for aiming
+	local s_Gunship = g_GameDirector:GetGunship(p_TeamId)
+	if false and s_Gunship then -- TODO: Remove "false", once the aiming works
+		local s_SeatsLeft = false
+		for i = 1, s_Gunship.entryCount - 1 do
+			if s_Gunship:GetPlayerInEntry(i) == nil then
+				s_SeatsLeft = true
+				break
+			end
+		end
+		if s_SeatsLeft then
+			return "SpawnInGunship"
+		end
 	end
 
 	-- CONQUEST
@@ -1458,7 +1513,7 @@ function BotSpawner:_GetUnlocks(p_Bot, p_TeamId, p_SquadId)
 			"persistence/unlocks/vehicles/artilleryairburst",
 		}
 		-- some variation in appearance
-		if MathUtils:GetRandomInt(1, 100) <= 50 then
+		if m_Utilities:CheckProbablity(50) then
 			table.insert(s_VehiclePerksToAdd, "persistence/unlocks/vehicles/mbtreactivearmor")
 			table.insert(s_VehiclePerksToAdd, "persistence/unlocks/vehicles/lbtreactivearmor")
 			table.insert(s_VehiclePerksToAdd, "persistence/unlocks/vehicles/ifvreloadupgrade")
@@ -1491,7 +1546,7 @@ function BotSpawner:_GetUnlocks(p_Bot, p_TeamId, p_SquadId)
 	-- Choose good available perk.
 	for _, l_PerkName in pairs(s_PossiblePerks) do
 		s_SelectedPerk = l_PerkName
-		if MathUtils:GetRandomInt(1, 100) <= 80 then -- Use the best available perk with this percentage.
+		if m_Utilities:CheckProbablity(80) then -- Use the best available perk with this percentage.
 			break
 		end
 	end
@@ -1506,7 +1561,8 @@ function BotSpawner:_GetUnlocks(p_Bot, p_TeamId, p_SquadId)
 			end
 		else
 			-- Vehicle perk.
-			for l_IndexVehiclePerk, l_VehiclePerkName in pairs(s_VehiclePerksToAdd) do
+			for l_IndexVehiclePerk = #s_VehiclePerksToAdd, 1, -1 do
+				local l_VehiclePerkName = s_VehiclePerksToAdd[l_IndexVehiclePerk]
 				if l_PerkName == l_VehiclePerkName then
 					table.remove(s_VehiclePerksToAdd, l_IndexVehiclePerk)
 					table.insert(s_Unlocks, s_CurrentUnlocks[l_Index])
@@ -1935,7 +1991,9 @@ function BotSpawner:_SwitchTeams()
 				s_NewTeam = Globals.NrOfTeams
 			end
 
-			l_Player.teamId = s_NewTeam
+			if not Config.BotTeamNames or not m_Utilities:isBot(l_Player) then
+				l_Player.teamId = s_NewTeam
+			end
 		end
 	end
 end