+++ ext/Shared/Settings/UpdateFlag.lua
@@ -7,5 +7,6 @@ UpdateFlag = {
 	YawPerSec = 4,
 	Language = 5,
 	MaxBots = 6,
-	Skill = 7
+	Skill = 7,
+	BotNames = 8
 }